make GC and verbose SSH optional with default true false
We add an extra option to control whether a GC is performed or not with
a default value of `true` and also an option to specify whether SSH is
run in verbose mode or not with a default of `false`. The former is
useful in situations where you want to be manually in control of GC and
the letter improves the log output in during normal use.